requirements.txt
flask
torch
transformers
sentencepiece
pytesseract
Pillow
gunicorn
